+++ frontend/package.json
@@ -10,10 +10,7 @@
     "@testing-library/user-event": "^7.2.1",
     "react": "^16.14.0",
     "react-dom": "^16.14.0",
-    "react-redux": "7.1.3",
-    "react-scripts": "^4.0.3",
-    "redux": "4.0.5",
-    "redux-thunk": "2.3.0"
+    "react-scripts": "^4.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
