--- frontend/src/constants/action-types.js
@@ -1,6 +1,0 @@
-export const DATA_LOAD_SUCCESS = 'DATA_LOAD_SUCCESS';
-export const REQUEST_DATA_LOAD = 'REQUEST_DATA_LOAD';
-export const SEARCH_STATIONS = 'SEARCH_STATIONS';
-export const FILTER_ELEC = 'FILTER_ELEC';
-export const FILTER_ELEC_FREE = 'FILTER_ELEC_FREE';
-export const FILTER_DOCK = 'FILTER_DOCK';
