--- frontend/src/index.css
@@ -1,108 +1,0 @@
-html, body {
-    font-size: 100%;
-}
-
-#ebike-no-classic {
-    margin: 1em 0;
-    padding: 1em 0;
-    border: black 1px dotted;
-}
-
-.panel {
-    text-align: center;
-}
-
-.panel input[type = "checkbox"] {-ms-transform: scale(2); /* IE */
-    -moz-transform: scale(2);
-    -webkit-transform: scale(2);
-    -o-transform: scale(2);
-    transform: scale(2);
-    padding: 10px;
-}
-
-.search {
-    font-weight: bold;
-    margin: 0.5em;
-    display: inline-block;
-}
-
-.checkbox-text {
-    margin-right: 1em;
-}
-
-h1, h2, h3, h4, h5, h6 {
-    margin: 0;
-    padding: 0;
-}
-
-h1 {
-	font-size: 3rem;
-	font-weight: 700;
-}
-
-h2 {
-    font-style: italic;
-    font-size: 1rem;
-    margin: 0.25em 0 1em 0;
-}
-
-.container-station {
-    border-top: 1px solid rgba(0,0,0,0.25);
-    margin: 1em 0;
-    padding: 1em 0;
-}
-
-h3.station-name {
-    font-weight: 700;
-}
-
-.infoSection {
-    margin: 0;
-    width: 29vw;
-    display: inline-block;
-    padding: 0;
-    text-align: center;
-}
-
-h4 {
-    font-weight: 700;
-    font-size: 3rem;
-}
-
-h5.descriptor {
-    font-weight: 500;
-}
-
-#container-footer {
-    font-size: 1.25em;
-    width: 100vw;
-    text-align: center;
-    position: fixed;
-    background-color: white;
-    bottom: 0;
-    display: block;
-    height: 10vh;
-    z-index: 10;
-}
-
-#content-footer {
-    position: relative;
-    top: 50%;
-    transform: translateY(-50%);
-}
-
-.searchOn .stationOff, .filterElecOn .elecOff, .filterDockOn .dockOff, .filterElecFreeOn .elecFreeOff {
-    display: none;
-}
-
-.btn {
-    border: #808080 1px solid;
-    display: inline-block;
-    padding: .2em .25em;
-    border-radius: .25em;
-}
-
-.btn-icon {
-    height: auto;
-    width: 1em;
-}
